remove NetVisualizer base class
